Type of bit indeces changed to uint32

+++ internal/filesystem/filesystem.go
@@ -84,11 +84,11 @@ func (fs FileSystem) CreateFirstFile(name string) error {
 	}
 	fs.Superblock.FreeInodeCount--
 
-	fileInode := inode.NewInode(true, 777, 0, 0, []uint32{uint32(blockIndex)})
-	inodeOffset := fs.GetInodeTableOffset() + int(fs.Superblock.InodeSize)*inodeIndex
+	fileInode := inode.NewInode(true, 777, 0, 0, []uint32{blockIndex})
+	inodeOffset := fs.GetInodeTableOffset() + int(fs.Superblock.InodeSize*inodeIndex)
 	fileInode.WriteAt(fs.dataFile, inodeOffset)
 
-	fs.currentDirectory.AddFile(uint32(inodeIndex), name)
+	fs.currentDirectory.AddFile(inodeIndex, name)
 
 	return nil
 }
